requirements.txt
numpy==1.26.4
pandas==2.2.2
scipy==1.11.4
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
optuna==3.6.1
shap==0.44.1
joblib==1.4.2
imbalanced-learn==0.12.2
gradio==4.36.1
google-generativeai==0.7.2
timm==0.9.16
kaggle==2.0.0
torch==2.2.2
torchvision==0.17.2
Pillow==10.3.0